following through with the latest design idiom with the other files
