+++ src/vtp/accept_ballot.py
@@ -23,11 +23,64 @@ See './accept_ballot.py -h' for usage information.
 
 """
 
-# pylint: disable=wrong-import-position   # import statements not top of file
+import argparse
 import sys
 
 from vtp.ops.accept_ballot_operation import AcceptBallotOperation
-from vtp.utils.election_config import ElectionConfig
+from vtp.utils.address import Address
+
+
+################
+# arg parsing
+################
+# pylint: disable=duplicate-code
+def parse_arguments():
+    """Parse arguments from a command line"""
+
+    parser = argparse.ArgumentParser(
+        description="""Will run the git based workflow on a VTP
+                    scanner node to accept the json rendering of the
+                    cast vote record of a voter's ballot.  The json
+                    file is read, the contests are extraced and
+                    submitted to separate git branches, one per
+                    contest, and pushed back to the Voter Center's VTP
+                    remote.
+
+                    In addition a voter's ballot receipt and offset
+                    are optionally printed.
+
+                    Either the location of the ballot_file or the
+                    associated address is required.
+                    """
+    )
+
+    Address.add_address_args(parser, True)
+    parser.add_argument(
+        "-m",
+        "--merge_contests",
+        action="store_true",
+        help="Will immediately merge the ballot contests (to master)",
+    )
+    parser.add_argument(
+        "--cast_ballot",
+        help="overrides an address - specifies a specific cast ballot",
+    )
+    parser.add_argument(
+        "-v",
+        "--verbosity",
+        type=int,
+        default=3,
+        help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
+    )
+    parser.add_argument(
+        "-n",
+        "--printonly",
+        action="store_true",
+        help="will printonly and not write to disk (def=True)",
+    )
+
+    return parser.parse_args()
+
 
 
 # pylint: disable=duplicate-code
@@ -39,15 +92,9 @@ def main():
     directory tree), and calls its main function.
     """
 
-    # Parse args first (ZZZ note logging interface)
-    _main = AcceptBallotOperation(sys.argv[1:])
-
-    # Create an VTP election config object
-    the_election_config = ElectionConfig()
-    the_election_config.parse_configs()
-
     # do it
-    _main.main(the_election_config)
+    abo = AcceptBallotOperation(parse_arguments())
+    abo.run()
 
 
 # If called directly via this file
