+++ src/vtp/generate_all_blank_ballots.py
@@ -26,7 +26,6 @@ See ../../docs/tech/executable-overview.md for the context in which this file wa
 """
 
 # Standard imports
-# pylint: disable=wrong-import-position   # import statements not top of file
 import argparse
 import logging
 import os
