+++ src/vtp/merge_contests.py
@@ -25,11 +25,67 @@ See 'merge_contests.py -h' for usage information.
 """
 
 # Standard imports
-# pylint: disable=wrong-import-position   # import statements not top of file
-import sys
+import argparse
 
 from vtp.ops.merge_contests_operation import MergeContestsOperation
-from vtp.utils.election_config import ElectionConfig
+
+
+################
+# arg parsing
+################
+# pylint: disable=duplicate-code
+def parse_arguments():
+    """Parse arguments from a command line"""
+
+    parser = argparse.ArgumentParser(
+        description="""Will run the git based workflow on a VTP
+    server node so to merge pending CVR contest branches into the
+    master git branch.
+
+    If there are less then the prerequisite number of already cast
+    contests, a warning will be printed/logged but no error will be
+    raised.  Supplying -f will flush all remaining contests to the
+    master branch.
+    """
+    )
+
+    parser.add_argument(
+        "-b", "--branch", default="", help="specify a specific branch to merge"
+    )
+    parser.add_argument(
+        "-m",
+        "--minimum_cast_cache",
+        type=int,
+        default=100,
+        help="the minimum number of cast ballots required prior to merging (def=100)",
+    )
+    parser.add_argument(
+        "-f",
+        "--flush",
+        action="store_true",
+        help="will flush the remaining unmerged contest branches",
+    )
+    parser.add_argument(
+        "-r",
+        "--remote",
+        action="store_true",
+        help="will merge remote branches instead of local branches",
+    )
+    parser.add_argument(
+        "-v",
+        "--verbosity",
+        type=int,
+        default=3,
+        help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
+    )
+    parser.add_argument(
+        "-n",
+        "--printonly",
+        action="store_true",
+        help="will printonly and not write to disk (def=True)",
+    )
+
+    return parser.parse_args()
 
 
 # pylint: disable=duplicate-code
@@ -41,15 +97,9 @@ def main():
     directory tree), and calls its main function.
     """
 
-    # Parse args first (ZZZ note logging interface)
-    _main = MergeContestsOperation(sys.argv[1:])
-
-    # Create an VTP election config object
-    the_election_config = ElectionConfig()
-    the_election_config.parse_configs()
-
     # do it
-    _main.main(the_election_config)
+    mco = MergeContestsOperation(parse_arguments())
+    mco.run()
 
 
 # If called directly via this file
