+++ src/vtp/ops/tally_contests_operation.py
@@ -24,113 +24,37 @@ a election.
 See 'tally_contests.py -h' for usage information.
 """
 
-# pylint: disable=wrong-import-position   # import statements not top of file
 # Standard imports
-import argparse
 import logging
-import re
-import sys
 
 # Local imports
 from vtp.utils.ballot import Ballot
-from vtp.utils.common import Shellout
+from vtp.utils.common import Common, Shellout
 from vtp.utils.contest import Tally
 from vtp.utils.exceptions import TallyException
+from vtp.utils.election_config import ElectionConfig
 
 
+# pylint: disable=too-few-public-methods
 class TallyContestsOperation:
     """A class to wrap the tally_contests.py script."""
 
-    def __init__(self, argv):
+    def __init__(self, parsed_args):
         """Only to module-ize the scripts and keep things simple and idiomatic."""
-        self.argv = argv
-        self.parsed_args = None
-        self.parse_arguments()
-
-    def __str__(self):
-        """Boilerplate"""
-        return "argv=" + str(self.argv) + "\n" + "parsed_args=" + str(self.parsed_args)
-
-    # pylint: disable=duplicate-code
-    def parse_arguments(self):
-        """Parse arguments from a command line"""
-
-        parser = argparse.ArgumentParser(
-            description="""Will tally all the contests so far merged to
-        the master branch and report the results.  The results are
-        computed on a voting center basis (git submodule) basis.
-
-        Note - the current implementation relies on git submodules
-        (individual git repos) to break up the tally data of an election.
-        If there is only one git repository and the election is large,
-        then a potentiallu large amount of memory will be used in
-        executing the tallies.  One short term fix for this is to limit
-        the number of contests being tallied.
-
-        Also note that the current implementation does not yet support
-        tallying across git submodules/repos.
-        """
-        )
-
-        parser.add_argument(
-            "-c",
-            "--contest_uid",
-            default="",
-            help="limit the tally to a specific contest uid",
-        )
-        parser.add_argument(
-            "-t",
-            "--track_contests",
-            default="",
-            help="a comma separated list of contests checks to track",
-        )
-        parser.add_argument(
-            "-x",
-            "--do_not_pull",
-            action="store_true",
-            help="Before tallying the votes, pull the ElectionData repo",
-        )
-        parser.add_argument(
-            "-v",
-            "--verbosity",
-            type=int,
-            default=3,
-            help="0 critical, 1 error, 2 warning, 3 info, 4 debug (def=3)",
-        )
-        #    parser.add_argument("-n", "--printonly", action="store_true",
-        #                            help="will printonly and not write to disk (def=True)")
-
-        self.parsed_args = parser.parse_args([str(x) for x in self.argv])
-        verbose = {
-            0: logging.CRITICAL,
-            1: logging.ERROR,
-            2: logging.WARNING,
-            3: logging.INFO,
-            4: logging.DEBUG,
-        }
-        logging.basicConfig(
-            format="%(message)s",
-            level=verbose[self.parsed_args.verbosity],
-            stream=sys.stdout,
-        )
-
-        # Validate required args
-        if self.parsed_args.track_contests:
-            if not bool(re.match("^[0-9a-f,]", self.parsed_args.track_contests)):
-                raise ValueError(
-                    "The track_contests parameter only accepts a comma separated (no spaces) "
-                    "list of contest checks/digests to track."
-                )
-            self.parsed_args.track_contests = self.parsed_args.track_contests.split(",")
-        else:
-            self.parsed_args.track_contests = []
+        self.parsed_args = parsed_args
 
     ################
     # main
     ################
     # pylint: disable=duplicate-code
-    def main(self, the_election_config):
+    def run(self):
         """Main function - see -h for more info"""
+
+        # Configure logging
+        Common.configure_logging(self.parsed_args.verbosity)
+
+        # Create a VTP ElectionData object if one does not already exist
+        the_election_config = ElectionConfig.configure_election()
 
         # git pull the ElectionData repo so to get the latest set of
         # remote CVRs branches
